feat: Added presets support to pull/push components
