+++ src/tasks/push-components.js
@@ -1,6 +1,7 @@
 const axios = require('axios')
 const fs = require('fs')
 const chalk = require('chalk')
+const PresetsLib = require('../utils/presets-lib')
 
 const isUrl = source => source.indexOf('http') === 0
 
@@ -53,65 +54,15 @@ module.exports = (api, options) => {
   return Promise.reject(new Error('Can not push invalid json - please provide a valid json file'))
 }
 
-const getComponentPresets = (component = {}, presets) => {
-  console.log(`${chalk.green('-')} Get presets from component ${component.name}`)
-
-  return presets.filter(preset => {
-    return preset.component_id === component.id
-  })
-}
-
-const filterPresetsFromTargetComponent = (presets, targetPresets) => {
-  console.log(chalk.blue('-') + ' Checking target presets to sync...')
-  const targetPresetsNames = targetPresets.map(preset => preset.name)
-  const newPresets = presets.filter(preset => !targetPresetsNames.includes(preset.name))
-  const updatePresetsSource = presets.filter(preset => targetPresetsNames.includes(preset.name))
-  const updatePresets = updatePresetsSource.map(source => {
-    const target = targetPresets.find(target => target.name === source.name)
-    return Object.assign({}, source, target, { image: source.image })
-  })
-
-  return {
-    newPresets,
-    updatePresets
-  }
-}
-
-const createPresets = async (api, presets = [], componentId, method = 'post') => {
-  const presetsSize = presets.length
-  console.log(`${chalk.blue('-')} Pushing ${presetsSize} ${method === 'post' ? 'new' : 'existing'} presets`)
-
-  try {
-    for (let i = 0; i < presetsSize; i++) {
-      const presetData = presets[i]
-      const presetId = method === 'put' ? `/${presetData.id}` : ''
-
-      await api[method](`presets${presetId}`, {
-        preset: {
-          name: presetData.name,
-          component_id: componentId,
-          preset: presetData.preset,
-          image: presetData.image
-        }
-      })
-    }
-
-    console.log(`${chalk.green('✓')} ${presetsSize} presets sync`)
-  } catch (e) {
-    console.error('An error ocurred when save the presets' + e.message)
-
-    return Promise.reject(e)
-  }
-}
-
 const push = async (api, components, presets) => {
+  const presetsLib = new PresetsLib({ oauthToken: api.accessToken, targetSpaceId: api.spaceId })
   try {
     const componentGroups = await api.getComponentGroups()
 
     const apiComponents = await api.getComponents()
 
     for (var i = 0; i < components.length; i++) {
-      const componentPresets = getComponentPresets(components[i], presets)
+      const componentPresets = presetsLib.getComponentPresets(components[i], presets)
       delete components[i].id
       delete components[i].created_at
 
@@ -157,15 +108,15 @@ const push = async (api, components, presets) => {
           console.error(`${chalk.red('X')} An error occurred when update component ${name}`)
           console.error(e.message)
         }
-        const presetsToSave = filterPresetsFromTargetComponent(
+        const presetsToSave = presetsLib.filterPresetsFromTargetComponent(
           componentPresets || [],
           componentTarget.data.component.all_presets || []
         )
         if (presetsToSave.newPresets.length) {
-          await createPresets(api, presetsToSave.newPresets, componentTarget.data.component.id, 'post')
+          await presetsLib.createPresets(presetsToSave.newPresets, componentTarget.data.component.id, 'post')
         }
         if (presetsToSave.updatePresets.length) {
-          await createPresets(api, presetsToSave.updatePresets, componentTarget.data.component.id, 'put')
+          await presetsLib.createPresets(presetsToSave.updatePresets, componentTarget.data.component.id, 'put')
         }
       } else {
         const { name } = components[i]
@@ -175,7 +126,7 @@ const push = async (api, components, presets) => {
             component: components[i]
           })
           if (componentPresets) {
-            createPresets(api, componentPresets, componentRes.data.component.id)
+            presetsLib.createPresets(componentPresets, componentRes.data.component.id)
           }
           console.log(`${chalk.green('✓')} Component ${name} has been updated in Storyblok!`)
         } catch (e) {
